test_main.py
import unittest
from unittest.mock import patch, Mock
import main

class TestSelectedFunction(unittest.TestCase):

    def test_selected_function_with_valid_input(self):
        # Test with valid input
        pass

    def test_selected_function_with_invalid_input(self):
        # Test with invalid input
        pass

    def test_selected_function_with_edge_cases(self):
        # Test with edge cases
        pass

    @patch('main.dependency_function')
    def test_selected_function_with_mocked_dependency(self, mock_dependency):
        # Test with mocked dependency
        mock_dependency.return_value = expected_value
        result = main.selected_function(input_value)
        self.assertEqual(result, expected_result)

if __name__ == '__main__':
    unittest.main()
tests/test_main.py
class TestMainWindow(unittest.TestCase):

    @patch('tkinter.Tk')
    def test_root_window_creation(self, mock_tk):
        main.root = mock_tk.return_value
        main.root.geometry.assert_called_with("300x300")
        main.root.title.assert_called_with("ProjectMohamedFarez")

    @patch('tkinter.Label')
    def test_label_creation(self, mock_label):
        main.root = Mock()
        main.Label(main.root, text=" Convert VIDEO TO AUDIO", bg='Lightblue', font=('Arial 13'))
        mock_label.assert_called_with(main.root, text=" Convert VIDEO TO AUDIO", bg='Lightblue', font=('Arial 13'))
        mock_label.return_value.pack.assert_called_once()

    @patch('tkinter.Button')
    def test_button_creation(self, mock_button):
        main.root = Mock()
        main.convert = Mock()
        main.Button(main.root, text=" Video to Audio", command=main.convert)
        mock_button.assert_called_with(main.root, text=" Video to Audio", command=main.convert)
        mock_button.return_value.pack.assert_called_once_with(pady=20)

    def test_convert_function(self):
        # Test cases for the convert function
        pass
tests/test_main.py
class TestConvertFunction(unittest.TestCase):

    @patch('tkinter.filedialog.askopenfilename')
    @patch('tkinter.filedialog.asksaveasfilename')
    @patch('moviepy.editor.VideoFileClip')
    def test_convert_with_valid_video_and_audio_files(self, mock_video_clip, mock_save_file, mock_open_file):
        mock_open_file.return_value = 'test_video.mp4'
        mock_save_file.return_value = 'test_audio.mp3'
        mock_video_clip.return_value = Mock(audio=Mock())

        convert()

        mock_open_file.assert_called_once()
        mock_video_clip.assert_called_once_with('test_video.mp4')
        mock_save_file.assert_called_once_with(defaultextension='.mp3', filetypes=[('Audio Files', '*.mp3')])
        mock_video_clip.return_value.audio.write_audiofile.assert_called_once_with('test_audio.mp3')

    @patch('tkinter.filedialog.askopenfilename')
    @patch('tkinter.filedialog.asksaveasfilename')
    def test_convert_with_no_video_file_selected(self, mock_save_file, mock_open_file):
        mock_open_file.return_value = ''

        convert()

        mock_open_file.assert_called_once()
        mock_save_file.assert_not_called()

    @patch('tkinter.filedialog.askopenfilename')
    @patch('tkinter.filedialog.asksaveasfilename')
    @patch('moviepy.editor.VideoFileClip')
    def test_convert_with_no_audio_file_selected(self, mock_video_clip, mock_save_file, mock_open_file):
        mock_open_file.return_value = 'test_video.mp4'
        mock_save_file.return_value = ''
        mock_video_clip.return_value = Mock(audio=Mock())

        convert()

        mock_open_file.assert_called_once()
        mock_video_clip.assert_called_once_with('test_video.mp4')
        mock_save_file.assert_called_once_with(defaultextension='.mp3', filetypes=[('Audio Files', '*.mp3')])
        mock_video_clip.return_value.audio.write_audiofile.assert_not_called()

    @patch('tkinter.messagebox.showinfo')
    @patch('tkinter.filedialog.askopenfilename')
    @patch('tkinter.filedialog.asksaveasfilename')
    @patch('moviepy.editor.VideoFileClip')
    def test_convert_with_successful_conversion(self, mock_video_clip, mock_save_file, mock_open_file, mock_show_info):
        mock_open_file.return_value = 'test_video.mp4'
        mock_save_file.return_value = 'test_audio.mp3'
        mock_video_clip.return_value = Mock(audio=Mock())

        convert()

        mock_show_info.assert_called_once_with('Congrats', 'Audio file is saved!')
